Don't export Date* functions via package root
Help to minimize overall package size for consumers not using Date*
components
