+++ packages/components/src/DateTimeFormat/DateTimeFormat.test.tsx
@@ -26,21 +26,11 @@
 import { assertSnapshot } from '@looker/components-test-utils'
 import React from 'react'
 import { Locales } from '../utils/i18n'
-import { DateFormat } from './DateFormat'
 import { DateTimeFormat } from './DateTimeFormat'
-import { TimeFormat } from './TimeFormat'
 const date = new Date('January 25, 1988 11:58:03')
 
 test('DateTime renders', () => {
   assertSnapshot(<DateTimeFormat>{date}</DateTimeFormat>)
-})
-
-test('DateFormat renders only date', () => {
-  assertSnapshot(<DateFormat>{date}</DateFormat>)
-})
-
-test('TimeFormat renders only time', () => {
-  assertSnapshot(<TimeFormat>{date}</TimeFormat>)
 })
 
 test('DateTimeFormat renders when passing specific locale', () => {
