+++ packages/components/src/DateTimeFormat/index.ts
@@ -24,6 +24,4 @@
 
  */
 
-export * from './DateFormat'
 export * from './DateTimeFormat'
-export * from './TimeFormat'
