+++ packages/components/src/FieldDateRange/FieldDateRange.tsx
@@ -26,13 +26,15 @@
 
 import React, { forwardRef, Ref } from 'react'
 import styled from 'styled-components'
-import { useID } from '../../../utils'
-import { useFormContext } from '../..'
+import { useID } from '../utils'
+import { useFormContext } from '../Form'
+import { InputDateRange, InputDateRangeProps } from '../InputDateRange'
 import {
-  InputDateRange,
-  InputDateRangeProps,
-} from '../../Inputs/InputDateRange'
-import { Field, FieldProps, omitFieldProps, pickFieldProps } from '../Field'
+  Field,
+  FieldProps,
+  omitFieldProps,
+  pickFieldProps,
+} from '../Form/Fields/Field'
 
 export interface FieldInputDateRangeProps
   extends FieldProps,
