+++ packages/components/src/InputDateRange/InputDateRange.test.tsx
@@ -27,8 +27,7 @@
 import React from 'react'
 import { fireEvent } from '@testing-library/react'
 import { renderWithTheme } from '@looker/components-test-utils'
-
-import { Locales } from '../../../utils'
+import { Locales } from '../utils/i18n'
 import { InputDateRange } from './InputDateRange'
 
 const realDateNow = Date.now.bind(global.Date)
