+++ packages/components/src/InputDateRange/InputDateRange.tsx
@@ -40,25 +40,20 @@ import max from 'lodash/max'
 import isEmpty from 'lodash/isEmpty'
 import isEqual from 'lodash/isEqual'
 import values from 'lodash/values'
-import { Icon } from '../../../Icon'
-import { ValidationType } from '../../ValidationMessage'
-
+import { Icon } from '../Icon'
+import { ValidationType } from '../Form/ValidationMessage'
 import {
   inputCSS,
   inputTextDisabled,
   inputTextHover,
   inputTextFocus,
   inputTextValidation,
-} from '../InputText'
-import { InlineInputTextBase } from '../InlineInputText'
-import { Calendar, CalendarLocalization } from '../../../Calendar'
-import {
-  Locales,
-  formatDateString,
-  parseDateFromString,
-  useID,
-  useReadOnlyWarn,
-} from '../../../utils'
+} from '../Form/Inputs/InputText'
+import { InlineInputTextBase } from '../Form/Inputs/InlineInputText'
+import { Calendar, CalendarLocalization } from '../Calendar'
+import { Locales, formatDateString, parseDateFromString } from '../utils/i18n'
+import { useID } from '../utils/useID'
+import { useReadOnlyWarn } from '../utils/useReadOnlyWarn'
 
 export interface InputDateRangeProps {
   value?: RangeModifier
