+++ packages/components/src/TimeFormat/TimeFormat.tsx
@@ -24,21 +24,9 @@
 
  */
 
-export * from './Checkbox'
-export * from './Combobox'
-export * from './InlineInputText'
-export * from './InlineTextArea'
-export * from './InputChips'
-export * from './InputColor'
-export * from './InputHidden'
-export * from './InputSearch'
-export * from './InputText'
-export * from './InputTime'
-export * from './InputTimeSelect'
-export * from './OptionsGroup'
-export * from './Radio'
-export * from './RangeSlider'
-export * from './Select'
-export * from './Slider'
-export * from './TextArea'
-export * from './ToggleSwitch'
+import React, { FC } from 'react'
+import { DateTimeFormatProps, DateTimeFormat } from '../DateTimeFormat'
+
+export const TimeFormat: FC<DateTimeFormatProps> = (props) => (
+  <DateTimeFormat {...props} date={false} />
+)
