+++ packages/components/src/utils/index.ts
@@ -39,7 +39,6 @@ export * from './usePopper'
 export * from './useScrollLock'
 export * from './useToggle'
 export * from './useWrapEvent'
-export * from './i18n'
 export * from './useMeasuredElement'
 export * from './useMouseDragPosition'
 export * from './usePreviousValue'
