+++ www/src/documentation/components/content/demos/DateFormatTable.tsx
@@ -24,7 +24,6 @@
 
  */
 import {
-  DateFormat,
   Table,
   TableBody,
   TableDataCell,
@@ -32,6 +31,7 @@ import {
   TableHeaderCell,
   TableRow,
 } from '@looker/components'
+import { DateFormat } from '@looker/components/src/DateFormat'
 import React from 'react'
 
 export const DateFormatTable = () => {
