+++ www/src/documentation/components/content/demos/LocaleTable.tsx
@@ -27,7 +27,6 @@
 import React from 'react'
 import toPairs from 'lodash/toPairs'
 import {
-  Locales,
   Table,
   TableBody,
   TableDataCell,
@@ -35,6 +34,7 @@ import {
   TableHeaderCell,
   TableRow,
 } from '@looker/components'
+import { Locales } from '@looker/components/src/utils/i18n'
 
 export const LocaleTable = () => {
   return (
