+++ www/src/documentation/components/forms/fields.mdx
@@ -4,6 +4,7 @@ figma: file/2MG6DoSjk4IaLnjjFCnKFf/Forms
 github: Form/Fields
 ---
 
+import { MessageBar } from '@looker/components'
 import { Props } from '../../../Shared'
 
 The `Field` variants of our input components provide convenient shorthand for the composition of Label, Component, and Validation message. As this is a normal use case for all form inputs, we will do our best to ensure that every `Input...` has an equivalent `Field...`.
@@ -228,6 +229,8 @@ This component is composed of an InputColor and a Label. It also has a `ColorWhe
 
 ## FieldDate
 
+Import path: `@looker/components/lib/FieldDate`
+
 <Props of="FieldDate" />
 
 This component is composed of an `InputDate` and a `Label`.
@@ -252,7 +255,9 @@ This component is composed of an `InputDate` and a `Label`.
 
 ## FieldDateRange
 
-<Props of="FieldDate" />
+Import path: `@looker/components/lib/FieldDateRange`
+
+<Props of="FieldDateRange" />
 
 This component is composed of an `InputDateRange` and a `Label`.
