+++ www/src/documentation/components/forms/input-date-range.mdx
@@ -4,6 +4,15 @@ propsOf: InputDateRange
 github: Form/Inputs/InputDateRange/InputDateRange.tsx
 ---
 
+import { MessageBar } from '@looker/components'
+
+<MessageBar my="large" canDismiss={false}>
+  Note: InputDatRange must be imported from its specific location rather than
+  the package root.
+</MessageBar>
+
+Import path: `@looker/components/lib/InputDateRange`
+
 `InputDateRange` provides an interface to select to a date range, either through the calendar interface or text input boxes.
 
 ## Value
